fix: API unmarshaling for Yandex changes
- Label.Id string -> FlexString (number -> string)
- InvocInfo.ExecDurationMillis int -> FlexInt (string -> int)
- Playlist.Tracks.Id uint64 -> FlexString (string)
- Artist/Album/Playlist/Owner/LikeAlbum/PinData ids uint64 -> FlexUint64 (string numbers)
- Station.Restrictions.Language/Diversity PossibleValues struct -> PossibleValues slice with custom Unmarshal (array vs object)
- convert all call sites: ArtistPopularTracks, Album, PlaylistTracks with uint64()

Resolves log errors:
  cannot unmarshal number into Label.id
  cannot unmarshal string into Playlist.Tracks.id
  cannot unmarshal string into ExecDurationMillis
  cannot unmarshal string into Artist.id
  cannot unmarshal array into PossibleValues

+++ api/api.go
@@ -304,7 +304,7 @@ func NewClient(name, token string) (*YaMusicClient, error) {
 		return nil, err
 	}
 
-	client.userid = clientStatus.Account.Uid
+	client.userid = uint64(clientStatus.Account.Uid)
 	client.sessionid = nowTimestamp()
 
 	return client, nil
